# src/main/resources/static/rest-api-doc.yaml
openapi: 3.0.1
info:
  title: Gwas Catalog Rest API V2
  description:  <p><img width="10%" alt="gwas_icon" src="../GWAS_Catalog_circle_178x178.png"></p>
                <p>Welcome to the <a rel="noopener noreferrer" target="_blank" href="https://www.ebi.ac.uk/gwas">GWAS Catalog</a> Restful API.</p>
                <p>The GWAS RESTful API provides programmatic access to the GWAS Catalog, enables integration of the GWAS Catalog data into custom applications and services, allowing external bioinformatics developers to leverage the curated GWAS datasets, and create more extended solutions for the scientific community.</p>
                <p>GWAS RESTful API V2 has been released with various enhancements & improvements over GWAS RESTful API V1. V1 will be deprecated no later than May 2026.<p>
                <p>Note that this API allows access to the literature-curated top associations & metadata (the same data that is available via the <a rel="noopener noreferrer" target="_blank" href="https://www.ebi.ac.uk/gwas">GWAS Catalog Website</a>. A second API enabling access to full genome-side summary statistics is under development.</p>
  version: "2.0"
servers:
- url: https://wwwdev.ebi.ac.uk/gwas/beta/rest/api
  #url: http://localhost:8087/gwas/rest/api
  description: Generated server url
...
paths:
...
  /v2/unpublished-studies/{accession_id}/unpublished-ancestries:
    get:
      tags:
      - unpublished-ancestries
      operationId: getUnpublishedAncestries
      summary: A GET request to retrieve ancestries for an unpublished study
      parameters:
      - name: accession_id
        in: path
        description: "The study’s GWAS Catalog accession ID <br/> <br/><i> Example\
          \ </i> : GCST90623790"
        required: true
        schema:
          type: string
      - name: sort
        in: query
        description: Fields to sort by
        required: false
        schema:
          type: string
          enum:
            - sample_size
            - cases
            - controls
      - name: direction
        in: query
        description: "Sort direction : asc or desc"
        required: false
        schema:
          type: string
          enum:
            - asc
            - desc
      responses:
        "500":
          description: Internal Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errors'
        "404":
          description: Entity Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errors'
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/collection-unpublished-ancestries'
  /v2/unpublished-studies/{accession_id}/unpublished-ancestries/{ancestry_id}:
    get:
      tags:
      - unpublished-ancestries
      operationId: getUnpublishedAncestry
      summary: A GET request to retrieve a single ancestry for an unpublished study
      parameters:
      - name: accession_id
        in: path
        description: "The study’s GWAS Catalog accession ID <br/> <br/><i> Example\
          \ </i> : GCST90623790"
        required: true
        schema:
          type: string
      - name: ancestry_id
        in: path
        description: "primary identifier of Ancestry table <br/> <br/><i> Example\
          \ </i> : 169062056"
        required: true
        schema:
          type: string
      responses:
        "500":
          description: Internal Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errors'
        "404":
          description: Entity Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errors'
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/unpublished-ancestries'
...
  /v2/studies/{accession_id}/ancestries/{ancestry_id}:
    get:
      tags:
      - ancestries
      operationId: getAncestry
      summary: A GET request to retrieve a single ancestry for a study in Catalog
      parameters:
      - name: accession_id
        in: path
        description: "The study’s GWAS Catalog accession ID <br/> <br/><i> Example\
          \ </i> : GCST000854"
        required: true
        schema:
          type: string
      - name: ancestry_id
        in: path
        description: "primary identifier of ancestry table <br/> <br/><i> Example\
          \ </i> : 7529"
        required: true
        schema:
          type: string
      responses:
        "500":
          description: Internal Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errors'
        "404":
          description: Entity Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errors'
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ancestries'
...
components:
  schemas:
    link:
      type: object
      properties:
        rel:
          type: string
        href:
          type: string
        hreflang:
          type: string
        media:
          type: string
        title:
          type: string
        type:
          type: string
        deprecation:
          type: string
        profile:
          type: string
        name:
          type: string
...
    errors:
      type: object
      properties:
        error_code:
          type: integer
          format: int32
          description: Error code
          example: "ORA-500001"
        error:
          type: string
          description: Error header
          example: "SQL Exception"
        error_message:
          type: string
          description: Error description
        service_path:
          type: string
          description: Endpoint details
          example: "/studies"
        timestamp:
          type: string
          description: Error timestamp
          example: 2021-10-19T22:22:55.015+0000
        report:
          type: object
          additionalProperties:
            type: string
            description: Addition attributes for the error
    ancestries:
      type: object
      properties:
        type:
          type: string
          description: Stage of the experimental design
          example: initial|replication
        number_of_individuals:
          type: integer
          format: int32
          description: Number of individuals in this group
          example: 10000
        ancestral_groups:
          type: array
          items:
            $ref: '#/components/schemas/ancestry-groups'
          description: Ancestry category group label
          example: European
        country_of_origin:
          type: array
          items:
            $ref: '#/components/schemas/countries'
          description: Country of origin of the individuals in the sample
          example: China
        country_of_recruitment:
          type: array
          items:
            $ref: '#/components/schemas/countries'
          description: List of country/countries where samples were recruited
          example: Japan
        links:
          type: array
          items:
            $ref: '#/components/schemas/link'
...
    ancestry-groups:
      type: object
      properties:
        ancestral_group:
          type: string
          description: Ancestry category group label
          example: European
...
    countries:
      type: object
      properties:
        major_area:
          type: string
          description: continent of the country of recruitment
          example: Europe
        region:
          type: string
          description: continent of the country of recruitment
          example: North Europe
        country_name:
          type: string
          description : country of recruitment
          example: U.K.
...
    unpublished-ancestries:
      type: object
      properties:
        study_tag:
          type: string
          description: A unique free-text label for each genome-wide association study in the publication
          example: white
        stage:
          type: string
          description: Stage of the experimental design
          example: discovery, replication
        sample_size:
          type: integer
          format: int32
          description: Number of individuals in this group
          example: 3000
        cases:
          type: integer
          format: int32
          description: Number of cases in this group
          example: 2000
        controls:
          type: integer
          format: int32
          description: Number of controls in this group
          example: 3000
        sample_description:
          type: string
          description: Additional  relevant information required for the interpretation of results
          example:  sex (males/females), age (adults/children)
        ancestry_category:
          type: string
          description: Broad ancestry category that best describes the sample
          example: European
        ancestry:
          type: string
          description: The most detailed ancestry descriptor
          example: All populations
        ancestry_description:
          type: string
          description: For founder or genetically isolated population, provide  description. If multiple founder/genetically isolated populations are included for the same ancestry category
        country_recruitment:
          type: string
          description: List of country/countries where samples were recruited
          example: U.S.|U.K.
        links:
          type: array
          items:
            $ref: '#/components/schemas/link'
    collection-unpublished-ancestries:
      type: object
      properties:
        links:
          type: array
          items:
            $ref: '#/components/schemas/link'
        content:
          type: array
          items:
            $ref: '#/components/schemas/unpublished-ancestries'
...